requirements.txt
django==1.11.26
djangorestframework==3.9.1